*Really* fix up next header.
